+++ weave/trace/serialize.py
@@ -1,9 +1,14 @@
-import json
 import typing
-import weakref
-from typing import Any, Generic, Iterator, Optional, Tuple, Union, ValuesView
+from collections import defaultdict
+from typing import (
+    Any,
+    DefaultDict,
+)
 
 from weave.trace import custom_objs
+from weave.trace.custom_weave_type_serialization_cache import (
+    CustomWeaveTypeSerializationCache,
+)
 from weave.trace.object_record import ObjectRecord
 from weave.trace.refs import ObjectRef, TableRef, parse_uri
 from weave.trace.serializer import get_serializer_for_obj
@@ -102,133 +107,17 @@ def from_json(obj: Any, project_id: str, server: TraceServerInterface) -> Any:
 # Importantly we can actually cache the results of both directions so that we
 # don't have to do the work more than once.
 
-K = typing.TypeVar("K")
-V = typing.TypeVar("V")
-
-
-class WeakKeyDictionarySupportingNonHashableKeys(Generic[K, V]):
-    def __init__(self) -> None:
-        self._id_to_data: dict[int, V] = {}
-        self._id_to_key: dict[int, K] = {}
-
-    def clear(self) -> None:
-        self._id_to_data.clear()
-        self._id_to_key.clear()
-
-    def get(self, key: K, default: Any = None) -> Union[V, Any]:
-        try:
-            return self.__getitem__(key)
-        except KeyError:
-            return default
-
-    def __getitem__(self, key: K) -> V:
-        item_id = id(key)
-        return self._id_to_data[item_id]
-
-    def __delitem__(self, key: K) -> None:
-        item_id = id(key)
-        if item_id in self._id_to_data:
-            del self._id_to_data[item_id]
-            del self._id_to_key[item_id]
-
-    def __setitem__(self, key: K, value: V) -> None:
-        item_id = id(key)
-        self._id_to_data[item_id] = value
-        self._id_to_key[item_id] = key
-        weakref.finalize(key, self._remove_item, item_id)
-
-    def _remove_item(self, item_id: int) -> None:
-        self._id_to_data.pop(item_id, None)
-        self._id_to_key.pop(item_id, None)
-
-    def __iter__(self) -> Iterator[K]:
-        return iter(self._id_to_key.values())
-
-    def __len__(self) -> int:
-        return len(self._id_to_data)
-
-    def keys(self) -> ValuesView[K]:
-        return self._id_to_key.values()
-
-    def values(self) -> ValuesView[V]:
-        return self._id_to_data.values()
-
-    def items(self) -> Iterator[Tuple[K, V]]:
-        return ((self._id_to_key[id], value) for id, value in self._id_to_data.items())
-
-
-class CustomWeaveTypeSerializationCache:
-    """Cache for custom weave type serialization.
-
-    Specifically, a dev can:
-    - store a serialization tuple of (deserialized object, serialized dict)
-    - retrieve the serialized dict for a deserialized object
-    - retrieve the deserialized object for a serialized dict
-
-    """
-
-    def __init__(self) -> None:
-        self._obj_to_dict: WeakKeyDictionarySupportingNonHashableKeys[Any, dict] = (
-            WeakKeyDictionarySupportingNonHashableKeys()
-        )
-        self._dict_to_obj: weakref.WeakValueDictionary[str, Any] = (
-            weakref.WeakValueDictionary()
-        )
-
-    def reset(self) -> None:
-        self._obj_to_dict.clear()
-        self._dict_to_obj.clear()
-
-    def store(self, obj: Any, serialized_dict: dict) -> None:
-        try:
-            self._store(obj, serialized_dict)
-        except Exception:
-            # Consider logging the exception here
-            pass
-
-    def _store(self, obj: Any, serialized_dict: dict) -> None:
-        self._obj_to_dict[obj] = serialized_dict
-        dict_key = self._get_dict_key(serialized_dict)
-        if dict_key is not None:
-            self._dict_to_obj[dict_key] = obj
-
-    def get_serialized_dict(self, obj: Any) -> Optional[dict]:
-        try:
-            return self._get_serialized_dict(obj)
-        except Exception:
-            # Consider logging the exception here
-            return None
-
-    def _get_serialized_dict(self, obj: Any) -> Optional[dict]:
-        return self._obj_to_dict.get(obj)
-
-    def get_deserialized_obj(self, serialized_dict: dict) -> Optional[Any]:
-        try:
-            return self._get_deserialized_obj(serialized_dict)
-        except Exception:
-            # Consider logging the exception here
-            return None
-
-    def _get_deserialized_obj(self, serialized_dict: dict) -> Optional[Any]:
-        dict_key = self._get_dict_key(serialized_dict)
-        return None if dict_key is None else self._dict_to_obj.get(dict_key)
-
-    def _get_dict_key(self, d: dict) -> Optional[str]:
-        try:
-            return json.dumps(d, sort_keys=True)
-        except Exception:
-            return None
-
-
 # Initialize the global cache
-_custom_weave_type_cache = CustomWeaveTypeSerializationCache()
+_custom_weave_type_cache_map: DefaultDict[str, CustomWeaveTypeSerializationCache] = (
+    defaultdict(CustomWeaveTypeSerializationCache)
+)
 
 
 def _to_json_custom_weave_type(
     obj: Any, project_id: str, server: TraceServerInterface
 ) -> dict:
     # Check if the object is already in the cache
-    cached_result = _custom_weave_type_cache.get_serialized_dict(obj)
+    cached_result = _custom_weave_type_cache_map[project_id].get_serialized_dict(obj)
     if cached_result is not None:
         return cached_result
 
@@ -254,7 +143,7 @@ def _to_json_custom_weave_type(
         result["load_op"] = load_op_uri
 
     # Store the result in the cache
-    _custom_weave_type_cache.store(obj, result)
+    _custom_weave_type_cache_map[project_id].store(obj, result)
     return result
 
 
@@ -262,7 +151,7 @@ def _from_json_custom_weave_type(
     obj: dict, project_id: str, server: TraceServerInterface
 ) -> Any:
     # Check if the serialized dict is already in the cache
-    cached_result = _custom_weave_type_cache.get_deserialized_obj(obj)
+    cached_result = _custom_weave_type_cache_map[project_id].get_deserialized_obj(obj)
     if cached_result is not None:
         return cached_result
 
@@ -270,5 +159,5 @@ def _from_json_custom_weave_type(
     result = custom_objs.decode_custom_obj(obj["weave_type"], files, obj.get("load_op"))
 
     # Store the result in the cache
-    _custom_weave_type_cache.store(result, obj)
+    _custom_weave_type_cache_map[project_id].store(result, obj)
     return result
